util: Move get_entity_name and get_entity_version into util-fabrics.
The get_entity_name and get_entity_version utility functions were only
used for fabrics.  Moves their declarations into private-fabrics.h and
their implementations into util-fabrics.c, and adds libnvmf
prefix for consistency with other fabrics-only utilities.
